requirements.txt
asgiref==3.7.2
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.5
django-cleanup==8.0.0
django-colorfield==0.11.0
django-cuser==2017.3.16
django-modeltranslation==0.18.11
django-rosetta==0.10.0
django-tinymce==3.6.1
django-widget-tweaks==1.5.0
idna==3.7
Pillow==10.0.1
polib==1.2.0
requests==2.32.3
rosetta==0.3
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.1
whitenoise==6.5.0
